requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
loguru==0.3.2
pkg-resources==0.0.0
PyYAML==5.1.2
requests==2.22.0
urllib3==1.25.3
